requirements.txt
#opencv-python==4.5.3.56
opencv-python
pyyaml
matplotlib
#numpy==1.19.3
numpy
pytest